pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ui-verdict"
version = "0.1.0"
description = "MCP server for agentic coding feedback loops — visual action-reaction verification"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pillow>=12.0.0",
    "opencv-python-headless>=4.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "mss>=9.0.0",
    "pynput>=1.7.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "wcag-contrast-ratio>=0.9",
]

[project.optional-dependencies]
macos = [
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Vision>=10.0",
    "pyobjc-framework-Quartz>=10.0",
]
vision = [
    "google-generativeai>=0.5.0",
    "anthropic>=0.30.0",
]

[project.scripts]
ui-verdict = "ui_verdict.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ui_verdict"]

[tool.pytest.ini_options]
testpaths = ["tests"]
